% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suspect.R
\name{check_suspect}
\alias{check_suspect}
\title{check_suspect}
\usage{
check_suspect(name)
}
\value{
A character vector
}
\description{
A function to check if the named suspect is the murderer/mastermind of the mystery"
}
\examples{
check_suspect("Reina Li")
}
