% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime_scene_report}
\alias{crime_scene_report}
\title{Data to showcase crime scene report}
\format{
A data frame with 1228 rows and 4 columns.
\describe{
\item{date}{The date when the crime took place}
\item{type}{The type of crime}
\item{description}{The description of the crime}
\item{city}{The city where the crime took place}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
crime_scene_report
}
\description{
A dataset containing the crime scene report.
}
\examples{
data(crime_scene_report)
}
\keyword{datasets}
