% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drivers_license}
\alias{drivers_license}
\title{Data to showcase drivers license}
\format{
A data frame with 10007 rows and 9 columns.
\describe{
\item{id}{Drivers license ID number}
\item{age}{Age of the driver}
\item{height}{Height of the driver}
\item{eye_color}{Eye color of the driver}
\item{hair_color}{Hair color of the driver}
\item{gender}{Gender of the driver}
\item{plate_number}{License plate number}
\item{car_make}{Make of the car}
\item{car_model}{Model of the car}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
drivers_license
}
\description{
A dataset containing drivers license information.
}
\examples{
data(drivers_license)
}
\keyword{datasets}
