% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facebook_event_checkin}
\alias{facebook_event_checkin}
\title{Data to showcase Facebook event check-ins}
\format{
A data frame with 20000 rows and 4 columns.
\describe{
\item{person_id}{Person ID of event attendee}
\item{event_id}{ID of the event}
\item{event_name}{Name of the event}
\item{date}{Date of the event}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
facebook_event_checkin
}
\description{
A dataset containing Facebook event check-ins.
}
\examples{
data(facebook_event_checkin)
}
\keyword{datasets}
