% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_fit_now_check_in}
\alias{get_fit_now_check_in}
\title{Data to showcase "Get Fit Now Gym" check-ins}
\format{
A data frame with 2703 rows and 4 columns.
\describe{
\item{membership_id}{ID of gym member}
\item{check_in_date}{Date of check-in}
\item{check_in_time}{Time of check-in}
\item{check_out_time}{Time of check-out}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
get_fit_now_check_in
}
\description{
A dataset containing "Get Fit Now Gym" check-ins.
}
\examples{
data(get_fit_now_check_in)
}
\keyword{datasets}
