% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_fit_now_member}
\alias{get_fit_now_member}
\title{Data to showcase "Get Fit Now" gym membership information}
\format{
A data frame with 184 rows and 5 columns.
\describe{
\item{id}{ID of gym member}
\item{person_id}{Person ID of gym member}
\item{name}{Name of the gym member}
\item{membership_start_date}{Date the gym member started the gym membership at "Get Fit Now"}
\item{membership_status}{Membership status of the gym member: regular, silver, gold}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
get_fit_now_member
}
\description{
A dataset containing "Get Fit Now" gym membership information.
}
\examples{
data(get_fit_now_member)
}
\keyword{datasets}
