% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Data to showcase income information}
\format{
A data frame with 7514 rows and 2 columns.
\describe{
\item{ssn}{Social Security Number of the person}
\item{annual_income}{Annual income of the person}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
income
}
\description{
A dataset containing income information.
}
\examples{
data(income)
}
\keyword{datasets}
