% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interview}
\alias{interview}
\title{Data to showcase interview transcripts}
\format{
A data frame with 4991 rows and 2 columns.
\describe{
\item{person_id}{Person ID}
\item{transcript}{Interview transcript provided}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
interview
}
\description{
A dataset containing interview transcripts.
}
\examples{
data(interview)
}
\keyword{datasets}
