% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{person}
\alias{person}
\title{Data to showcase personal information}
\format{
A data frame with 10011 rows and 6 columns.
\describe{
\item{id}{Person ID}
\item{name}{Name of the person}
\item{license_id}{Drivers license ID number of the person}
\item{address_number}{Address number of the person}
\item{address_street_name}{Address street name of the person}
\item{ssn}{Social Security Number of the person}
}
}
\source{
Adapted from "The SQL Murder Mystery" originally created by Joon Park and Cathy He while they were Knight Lab fellows.
}
\usage{
person
}
\description{
A dataset containing personal information.
}
\examples{
data(person)
}
\keyword{datasets}
